Return []float32 from GetEmb

+++ util.go
@@ -5,24 +5,24 @@ import (
 	"encoding/binary"
 )
 
-func vecToBytes(vec []float64, order binary.ByteOrder) []byte {
+func vecToBytes(vec []float32, order binary.ByteOrder) []byte {
 	buf := new(bytes.Buffer)
 	for _, v := range vec {
-		binary.Write(buf, order, float32(v))
+		binary.Write(buf, order, v)
 	}
 	return buf.Bytes()
 }
 
-func bytesToVec(data []byte, order binary.ByteOrder) ([]float64, error) {
+func bytesToVec(data []byte, order binary.ByteOrder) ([]float32, error) {
 	size := len(data) / 4
-	vec := make([]float64, size)
+	vec := make([]float32, size)
 	buf := bytes.NewReader(data)
 	var v float32
 	for i := range vec {
 		if err := binary.Read(buf, order, &v); err != nil {
 			return nil, err
 		}
-		vec[i] = float64(v)
+		vec[i] = v
 	}
 	return vec, nil
 }
